refactor(models): restructure FibexCore per spec packages - CoreCommunication pkg, Timing, FibexElement, CommConnectorPort, EcuInstance (Rule 0007 A)
- CoreCommunication.py -> CoreCommunication/ package; Timing.py moves under it
- FibexElement -> FibexCore/__init__.py (spec package FibexCore)
- CommConnectorPort, EcuInstance -> CoreTopology.py (spec package CoreTopology)
- Break CoreTopology <-> CoreCommunication cycle with a function-level
  LinScheduleTable import in createLinScheduleTable; CommunicationDirectionType
  stays TYPE_CHECKING (spec: CoreCommunication) and the get_type_hints test
  resolves it via localns
