+++ src/armodel/models/M2/AUTOSARTemplates/SystemTemplate/Fibex/FibexCore/CoreCommunication/__init__.py
@@ -2,38 +2,22 @@ from __future__ import annotations
 
 from abc import ABC
 from typing import List, Optional, TYPE_CHECKING
-from armodel.models.M2.AUTOSARTemplates.GenericStructure.GeneralTemplateClasses.Identifiable import ARElement, Identifiable, Describable, PackageableElement
+from armodel.models.M2.AUTOSARTemplates.GenericStructure.GeneralTemplateClasses.Identifiable import ARElement, Identifiable, Describable
 from armodel.models.M2.AUTOSARTemplates.GenericStructure.GeneralTemplateClasses.ArObject import ARObject
 from armodel.models.M2.AUTOSARTemplates.GenericStructure.GeneralTemplateClasses.PrimitiveTypes import AREnum, ARLiteral, ARNumerical, ARPositiveInteger, Boolean, ByteOrderEnum
 from armodel.models.M2.AUTOSARTemplates.GenericStructure.GeneralTemplateClasses.PrimitiveTypes import Integer, PositiveInteger, RefType, ARBoolean, String
 from armodel.models.M2.AUTOSARTemplates.GenericStructure.GeneralTemplateClasses.PrimitiveTypes import TimeValue, UnlimitedInteger
 from armodel.models.M2.MSR.DataDictionary.DataDefProperties import SwDataDefProps
 from armodel.models.M2.AUTOSARTemplates.CommonStructure.Filter import DataFilter
-from armodel.models.M2.AUTOSARTemplates.SystemTemplate.Fibex.FibexCore.Timing import TransmissionModeDeclaration, TriggerIPduSendCondition
+from armodel.models.M2.AUTOSARTemplates.SystemTemplate.Fibex.FibexCore.CoreCommunication.Timing import TransmissionModeDeclaration, TriggerIPduSendCondition
+from armodel.models.M2.AUTOSARTemplates.SystemTemplate.Fibex.FibexCore import FibexElement
+from armodel.models.M2.AUTOSARTemplates.SystemTemplate.Fibex.FibexCore.CoreTopology import CommConnectorPort
 
 if TYPE_CHECKING:
     from armodel.models.M2.AUTOSARTemplates.CommonStructure import ValueSpecification
     from armodel.models.M2.AUTOSARTemplates.SWComponentTemplate.Communication import HandleOutOfRangeEnum
     from armodel.models.M2.AUTOSARTemplates.SystemTemplate.DataMapping import DataTypePolicyEnum
     from armodel.models.M2.AUTOSARTemplates.SystemTemplate.Transformer import TransformationISignalProps
-
-
-class FibexElement(PackageableElement, ABC):
-    """
-    ASAM FIBEX elements specifying Communication and Topology.
-    """
-
-    # FibexElement method parity checklist:
-    # Spec: AUTOSAR_CP_TPS_SystemTemplate.pdf, Table F.64, p.2026
-    # Spec verified: R23-11
-    # Columns: impl / docstring / test / reader / writer   ([—] = no XML element)
-    # [x] __init__                     [x] impl  [x] docstring  [x] test  [—] reader  [—] writer
-
-    def __init__(self, parent: ARObject, short_name: str):
-        if type(self) is FibexElement:
-            raise TypeError("FibexElement is an abstract class.")
-
-        super().__init__(parent, short_name)
 
 
 class PduToFrameMapping(Identifiable):
@@ -2478,46 +2462,6 @@ class CommunicationDirectionType(AREnum):
 
     def __init__(self):
         super().__init__([CommunicationDirectionType.ENUM_IN, CommunicationDirectionType.ENUM_OUT])
-
-
-class CommConnectorPort(Identifiable, ABC):
-    """
-    The Ecu communication relationship defines which signals, Pdus and frames are actually received and transmitted by this ECU. For each signal, Pdu or Frame that is transmitted or received and used by the Ecu an association between an ISignalPort, IPduPort or FramePort with the corresponding Triggering shall be created. An ISignalPort shall be created only if the corresponding signal is handled by COM (RTE or Signal Gateway). If a Pdu Gateway ECU only routes the Pdu without being interested in the content only a FramePort and an IPduPort needs to be created.
-
-    [constr_9103] Existence of communicationDirection: For each CommConnectorPort, the attribute communicationDirection shall exist at the time when the System Description is complete.
-    """
-
-    # CommConnectorPort method parity checklist:
-    # Spec: AUTOSAR_CP_TPS_SystemTemplate.pdf, Table 6.1, p.303
-    # Spec verified: R23-11
-    # Columns: impl / docstring / test / reader / writer   ([—] = no XML element)
-    # [x] __init__                     [x] impl  [x] docstring  [x] test  [—] reader  [—] writer
-    # [x] getCommunicationDirection    [x] impl  [x] docstring  [x] test  [—] reader  [x] writer
-    # [x] setCommunicationDirection    [x] impl  [x] docstring  [x] test  [x] reader  [—] writer
-
-    def __init__(self, parent: ARObject, short_name: str):
-        if type(self) is CommConnectorPort:
-            raise TypeError("CommConnectorPort is an abstract class.")
-
-        super().__init__(parent, short_name)
-
-        # Communication Direction of the Connector Port (input or output Port).
-        self.communicationDirection: Optional[CommunicationDirectionType] = None
-
-    def getCommunicationDirection(self) -> Optional[CommunicationDirectionType]:
-        """
-        Communication Direction of the Connector Port (input or output Port).
-        """
-        return self.communicationDirection
-
-    def setCommunicationDirection(self, value: Optional[CommunicationDirectionType]) -> "CommConnectorPort":
-        """
-        Communication Direction of the Connector Port (input or output Port).
-        A None value is a no-op and does not overwrite an existing communicationDirection.
-        """
-        if value is not None:
-            self.communicationDirection = value
-        return self
 
 
 class FramePort(CommConnectorPort):
